+++ R/LearnerClassifLDA.R
@@ -29,18 +29,15 @@ LearnerClassifLDA = R6Class("LearnerClassifLDA",
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
-      ps = ParamSet$new(list(
-        ParamUty$new("prior", tags = "train"),
-        ParamDbl$new("tol", tags = "train"),
-        ParamFct$new("method", default = "moment", levels = c(
-          "moment", "mle", "mve", "t"), tags = "train"),
-        ParamInt$new("nu", tags = "train"),
-        ParamFct$new("predict.method",
-          default = "plug-in",
-          levels = c("plug-in", "predictive", "debiased"), tags = "predict"),
-        ParamUty$new("predict.prior", tags = "predict"),
-        ParamUty$new("dimen", tags = "predict")
-      ))
+      ps = ps(
+        dimen          = p_uty(tags = "predict"),
+        method         = p_fct(c("moment", "mle", "mve", "t"), default = "moment", tags = "train"),
+        nu             = p_int(tags = "train"),
+        predict.method = p_fct(c("plug-in", "predictive", "debiased"), default = "plug-in", tags = "predict"),
+        predict.prior  = p_uty(tags = "predict"),
+        prior          = p_uty(tags = "train"),
+        tol            = p_dbl(tags = "train")
+      )
       ps$add_dep("nu", "method", CondEqual$new("t"))
 
       super$initialize(
@@ -57,31 +54,22 @@ LearnerClassifLDA = R6Class("LearnerClassifLDA",
 
   private = list(
     .train = function(task) {
+      pv = self$param_set$get_values(tags = "train")
       formula = task$formula()
-      mlr3misc::invoke(MASS::lda, formula,
-        data = task$data(),
-        .args = self$param_set$get_values(tags = "train"))
+      invoke(MASS::lda, formula, data = task$data(), .args = pv)
     },
 
     .predict = function(task) {
-      pars = self$param_set$get_values(tags = "predict")
-      if (!is.null(pars$predict.method)) {
-        pars$method = pars$predict.method
-        pars$predict.method = NULL
-      }
-      if (!is.null(pars$predict.prior)) {
-        pars$prior = pars$predict.prior
-        pars$predict.prior = NULL
-      }
+      pv = self$param_set$get_values(tags = "predict")
+      pv = rename(pv, c("predict.method", "predict.prior"), c("method", "prior"))
       newdata = task$data(cols = task$feature_names)
-      p = mlr3misc::invoke(predict, self$model,
-        newdata = newdata,
-        .args = self$param_set$get_values(tags = "predict"))
+
+      p = invoke(predict, self$model, newdata = newdata, .args = pv)
 
       if (self$predict_type == "response") {
-        list(response = p$class)
+        list(response = p[["class"]])
       } else {
-        list(response = p$class, prob = p$posterior)
+        list(response = p[["class"]], prob = p[["posterior"]])
       }
     }
   )