+++ R/LearnerClassifLogReg.R
@@ -6,6 +6,11 @@
 #' Classification via logistic regression.
 #' Calls [stats::glm()] with `family` set to `"binomial"`.
 #'
+#' @section Internal Encoding:
+#' Starting with \CRANpkg{mlr3} v0.5.0, the order of class labels is reversed prior to
+#' model fitting to comply to the [stats::glm()] convention that the negative class is provided
+#' as the first factor level.
+#'
 #' @section Custom mlr3 defaults:
 #' - `model`:
 #'   - Actual default: `TRUE`.
@@ -28,56 +33,68 @@ LearnerClassifLogReg = R6Class("LearnerClassifLogReg",
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
-      ps = ParamSet$new(list(
-        ParamLgl$new("singular.ok", default = TRUE, tags = "train"),
-        ParamLgl$new("x", default = FALSE, tags = "train"),
-        ParamLgl$new("y", default = TRUE, tags = "train"),
-        ParamLgl$new("model", default = TRUE, tags = "train"),
-        ParamUty$new("etastart", tags = "train"),
-        ParamUty$new("mustart", tags = "train"),
-        ParamUty$new("start", default = NULL, tags = "train"),
-        ParamUty$new("offset", tags = "train"),
-        ParamDbl$new("epsilon", default = 1e-8, tags = c("train", "control")),
-        ParamDbl$new("maxit", default = 25, tags = c("train", "control")),
-        ParamLgl$new("trace", default = FALSE, tags = c("train", "control")),
-        ParamLgl$new("se.fit", default = FALSE, tags = "predict"),
-        ParamUty$new("dispersion", default = NULL, tags = "predict")
-      ))
+      ps = ps(
+        dispersion  = p_uty(default = NULL, tags = "predict"),
+        epsilon     = p_dbl(default = 1e-8, tags = c("train", "control")),
+        etastart    = p_uty(tags = "train"),
+        maxit       = p_dbl(default = 25, tags = c("train", "control")),
+        model       = p_lgl(default = TRUE, tags = "train"),
+        mustart     = p_uty(tags = "train"),
+        offset      = p_uty(tags = "train"),
+        singular.ok = p_lgl(default = TRUE, tags = "train"),
+        start       = p_uty(default = NULL, tags = "train"),
+        trace       = p_lgl(default = FALSE, tags = c("train", "control")),
+        x           = p_lgl(default = FALSE, tags = "train"),
+        y           = p_lgl(default = TRUE, tags = "train")
+      )
 
       super$initialize(
         id = "classif.log_reg",
         param_set = ps,
         predict_types = c("response", "prob"),
         feature_types = c("logical", "integer", "numeric", "character", "factor", "ordered"),
-        properties = c("weights", "twoclass"),
+        properties = c("weights", "twoclass", "loglik"),
         packages = "stats",
         man = "mlr3learners::mlr_learners_classif.log_reg"
       )
+    },
+
+    #' @description
+    #' Extract the log-likelihood (e.g., via [stats::logLik()] from the fitted model.
+    loglik = function() {
+      extract_loglik(self)
     }
   ),
 
   private = list(
     .train = function(task) {
-      pars = self$param_set$get_values(tags = "train")
+      pv = self$param_set$get_values(tags = "train")
       if ("weights" %in% task$properties) {
-        pars = insert_named(pars, list(weights = task$weights$weight))
+        pv = insert_named(pv, list(weights = task$weights$weight))
       }
 
-      mlr3misc::invoke(stats::glm,
-        formula = task$formula(), data = task$data(),
-        family = "binomial", model = FALSE, .args = pars, .opts = opts_default_contrasts)
+      # logreg expects the first label to be the negative class, contrary
+      # to the mlr3 convention that the positive class comes first.
+      tn = task$target_names
+      data = task$data()
+      data[[tn]] = swap_levels(data[[tn]])
+
+      invoke(stats::glm,
+        formula = task$formula(), data = data,
+        family = "binomial", model = FALSE, .args = pv, .opts = opts_default_contrasts)
     },
 
     .predict = function(task) {
+      pv = self$param_set$get_values(tags = "predict")
+      lvls = c(task$negative, task$positive)
       newdata = task$data(cols = task$feature_names)
 
-      p = unname(predict(self$model, newdata = newdata, type = "response"))
-      levs = levels(self$model$data[[task$target_names]])
+      p = unname(invoke(predict, object = self$model, newdata = newdata, type = "response", .args = pv))
 
       if (self$predict_type == "response") {
-        list(response = ifelse(p < 0.5, levs[1L], levs[2L]))
+        list(response = ifelse(p < 0.5, lvls[1L], lvls[2L]))
       } else {
-        list(prob = pvec2mat(p, levs))
+        list(prob = pvec2mat(p, lvls))
       }
     }
   )