+++ R/zzz.R
@@ -1,6 +1,7 @@
 #' @import data.table
 #' @import paradox
 #' @import mlr3misc
+#' @import checkmate
 #' @importFrom R6 R6Class
 #' @importFrom mlr3 mlr_learners LearnerClassif LearnerRegr
 #' @importFrom stats predict
@@ -47,15 +48,7 @@ register_mlr3 = function() {
 }
 
 .onLoad = function(libname, pkgname) { # nolint
-  register_mlr3()
-  setHook(packageEvent("mlr3", "onLoad"), function(...) register_mlr3(), action = "append")
-} # nocov end
-
-.onUnload = function(libpath) { # nolint
-  event = packageEvent("mlr3", "onLoad")
-  hooks = getHook(event)
-  pkgname = vapply(hooks, function(x) environment(x)$pkgname, NA_character_)
-  setHook(event, hooks[pkgname != "mlr3learners"], action = "replace")
+  register_namespace_callback(pkgname, "mlr3", register_mlr3)
 } # nocov end
 
 leanify_package()