+++ inst/paramtest/test_paramtest_classif.ranger.R
@@ -16,7 +16,8 @@ test_that("classif.ranger", {
     "quantreg", # regression only
     "dependent.variable.name", # handled via mlr3
     "status.variable.name", # handled via mlr3
-    "classification" # handled via mlr3
+    "classification", # handled via mlr3
+    "mtry.ratio" # custom hyperpar
   )
 
   ParamTest = run_paramtest(learner, fun, exclude)